+++ client/proposer/BUILD.bazel
@@ -20,6 +20,27 @@ go_test(
         "//shared/testutil:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
+    ],
+)
+
+# by default gazelle tries to add all the test files to the  first
+# go_test; we need to exclude the tests that we want to configure
+# in a particular way
+# gazelle:exclude service_norace_test.go
+
+go_test(
+    name = "go_norace_test",
+    srcs = ["service_norace_test.go"],
+    embed = [":go_default_library"],
+    race = "off",  # TODO(#377): fix issues with race detection testing.
+    deps = [
+        "//client/internal:go_default_library",
+        "//client/mainchain:go_default_library",
+        "//client/params:go_default_library",
+        "//client/syncer:go_default_library",
+        "//proto/sharding/p2p/v1:go_default_library",
+        "//shared/database:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
 )
